Split out resumable to module
Also adds a newtype wrapper around the resumable session so that we have
a bit of type safety.
