+++ src/error.rs
@@ -15,6 +15,8 @@ pub enum Error {
     InvalidPrefix(&'static str),
     #[error("Sequence {0} is not allowed")]
     InvalidSequence(&'static str),
+    #[error("Failed to parse URI")]
+    InvalidUri(UriError),
     #[error("HTTP error")]
     Http(#[source] HttpError),
     #[error("HTTP status")]
@@ -133,6 +135,29 @@ impl From<serde_json::Error> for Error {
 impl From<serde_urlencoded::ser::Error> for Error {
     fn from(e: serde_urlencoded::ser::Error) -> Self {
         Error::UrlEncode(e)
+    }
+}
+
+#[derive(Debug, thiserror::Error)]
+pub struct UriError(#[source] http::uri::InvalidUri);
+
+impl fmt::Display for UriError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+impl PartialEq for UriError {
+    fn eq(&self, other: &Self) -> bool {
+        // This is **TERRIBLE** but all of the error details are unnecessarily
+        // private and it doesn't implement PartialEq ARGH
+        self.0.to_string() == other.0.to_string()
+    }
+}
+
+impl From<http::uri::InvalidUri> for Error {
+    fn from(e: http::uri::InvalidUri) -> Self {
+        Error::InvalidUri(UriError(e))
     }
 }
